+++ OutSouringProject/src/main/java/com/sparta/outsouringproject/common/aop/OrderLogAop.java
@@ -19,15 +19,15 @@ public class OrderLogAop {
         Long storeId = orderCreateResponseDto.getStoreId();
         Long orderId = orderCreateResponseDto.getOrderId();
         LocalDateTime cur = LocalDateTime.now();
-        log.info("::: 신규 주문 :::\n가게 ID: {} / 주문 ID: {} / 요청시각: {}", cur, storeId, orderId);
+        log.info("::: 신규 주문 ::: 가게 ID: {} / 주문 ID: {} / 요청시각: {}", storeId, orderId, cur);
     }
 
     @AfterReturning("execution(* com.sparta.outsouringproject.order.service.OrderServiceImpl.changeOrderStatus(..))")
     public void logOrderStatusChange(JoinPoint joinPoint) {
         Object[] args = joinPoint.getArgs();
-        Long storeId = (Long) args[0]; // todo: @AuthUser 추가되면 1로 바꿔야함
-        Long orderId = (Long) args[1]; // todo: @AuthUser 추가되면 2로 바꿔야함
+        Long storeId = (Long) args[1]; // todo: @AuthUser 추가되면 1로 바꿔야함
+        Long orderId = (Long) args[2]; // todo: @AuthUser 추가되면 2로 바꿔야함
         LocalDateTime cur = LocalDateTime.now();
-        log.info("::: 배달상태 변경 :::\n가게 ID: {} / 주문 ID: {} / 요청시각: {}", cur, storeId, orderId);
+        log.info("::: 배달상태 변경 ::: 가게 ID: {} / 주문 ID: {} / 요청시각: {}", storeId, orderId, cur);
     }
 }
